requirements.txt
streamlit
bcrypt
opencv-python-headless
numpy
python-dotenv
pillow
pandas
plotly
streamlit-option-menu
streamlit-autorefresh
firebase-admin
pdf2image
pypdf2
google-generativeai
PyMuPDF
rapidfuzz
# insightface
mediapipe
onnxruntime
gdown
streamlit-webrtc
av
setuptools==68.2.2
